backend/gdelt-analysis-main/requirements.txt
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
db-dtypes>=1.4.0